+++ airline-data/src/main/scala/com/patson/model/media/WikimediaResource.scala
@@ -1,0 +1,7 @@
+package com.patson.model.media
+
+case class WikimediaResource(resourceId : Int, resourceType : ResourceType.Value, url : String, maxAgeDeadline : Option[Long])
+
+object ResourceType extends Enumeration {
+  val CITY_IMAGE, AIRPORT_IMAGE = Value
+}
